+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * NEW: disable select aura displays in the options
 * NEW: option to disable player frame while not in a party
 * FIX: party-/raidframes missing after using any test mode
+* FIX: characters without spec now don't make lua errors but get an error message
 ## v0.0.6a (2020/11/29)
 * NEW: add option to show buff/debuff tooltips
 * Added a lot of missing defensives. Please report if I missed some :)